# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Watson Orchestration Backend API",
    "version": "1.0.0",
    "description": "REST API for accessing YFS Statistics Detail data from PostgreSQL database for WatsonX Orchestrate integration",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "https://watson-orchestration-backend-24598467596.europe-west1.run.app",
      "description": "Development server"
    },
    {
      "url": "https://api.production.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/query/stats": {
      "get": {
        "summary": "Get General Statistics",
        "description": "Retrieve general statistics for a specified date range",
        "tags": ["Query"],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "date_start",
            "in": "query",
            "description": "Start date (YYYY-MM-DD) or relative date (e.g., 'last_7_days', 'yesterday')",
            "required": false,
            "schema": {
              "type": "string",
              "default": "last_7_days"
            }
          },
          {
            "name": "date_end",
            "in": "query",
            "description": "End date (YYYY-MM-DD) or relative date (e.g., 'today')",
            "required": false,
            "schema": {
              "type": "string",
              "default": "today"
            }
          },
          {
            "name": "service_name",
            "in": "query",
            "description": "Filter by service name (partial match, case-insensitive)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "service_type",
            "in": "query",
            "description": "Filter by service type (exact match)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["API", "INTEGRATION", "AGENT", "SERVICE"]
            }
          },
          {
            "name": "statistic_name",
            "in": "query",
            "description": "Filter by statistic name (partial match, case-insensitive)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/query/components/list": {
      "get": {
        "summary": "List Components (Optimized & Simplified)",
        "description": "Get a list of distinct servers with optional service details. By default, returns only server names for performance. Set `include_services=true` or `include_services=Y` to include full service details. **OPTIMIZED**: Single SQL query with conditional service inclusion.",
        "tags": ["Query"],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Search term to filter server names (case-insensitive, supports partial matching)",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "YFSAPP01"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of servers to return (default: 300)",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 300,
              "maximum": 300,
              "minimum": 1
            },
            "example": 10
          },
          {
            "name": "include_services",
            "in": "query",
            "description": "Include service details in response. Accepts 'true', 'Y' to include services. Default is false (N) for better performance.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["true", "false", "Y", "N"],
              "default": "false"
            },
            "example": "true"
          }
        ],
        "responses": {
          "200": {
            "description": "Components list retrieved successfully. Response format depends on `include_services` parameter:\n- **Default (include_services=false/N)**: Returns only server names for optimal performance\n- **With services (include_services=true/Y)**: Returns servers with full service details including statistic names",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComponentsListResponse"
                },
                "examples": {
                  "without_services": {
                    "summary": "Default response (include_services=false or N)",
                    "value": {
                      "success": true,
                      "data": [
                        {
                          "server_name": "YFSAPP01"
                        },
                        {
                          "server_name": "YFSAPP02"
                        },
                        {
                          "server_name": "YFSAPP03"
                        }
                      ],
                      "metadata": {
                        "searchTerm": null,
                        "count": 74,
                        "limit": 300,
                        "include_services": false
                      }
                    }
                  },
                  "with_services": {
                    "summary": "Full response (include_services=true or Y)",
                    "value": {
                      "success": true,
                      "data": [
                        {
                          "server_name": "YFSAPP01",
                          "services": [
                            {
                              "service_name": "processOrder",
                              "service_type": "INTEGRATION",
                              "context_name": "YFSExtnOrderProcessOrderImplService",
                              "statistic_names": ["Average", "Invocations", "Maximum", "Minimum"]
                            },
                            {
                              "service_name": "getOrderDetails",
                              "service_type": "INTEGRATION",
                              "context_name": "YFSExtnOrderProcessOrderImplService",
                              "statistic_names": ["Average", "Invocations", "Maximum"]
                            }
                          ]
                        },
                        {
                          "server_name": "YFSAPP02",
                          "services": [
                            {
                              "service_name": "updateInventory",
                              "service_type": "INTEGRATION",
                              "context_name": "YFSExtnInventoryManagementService",
                              "statistic_names": ["Average", "Invocations"]
                            }
                          ]
                        }
                      ],
                      "metadata": {
                        "searchTerm": null,
                        "count": 74,
                        "limit": 300,
                        "total_services": 180,
                        "include_services": true
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Failed to fetch component list"
                    }
                  }
                }
              }
            }
          }
        },
        "x-performance": {
          "optimization": "Single SQL query with ARRAY_AGG",
          "improvement": "75x faster (75 queries reduced to 1)",
          "technique": "CTE with GROUP BY and ARRAY_AGG for statistic names",
          "response_time": "< 300ms for 74 servers with distinct services",
          "data_structure": "Simplified - removed aggregated values, grouped statistics into arrays"
        }
      }
    },
    "/api/query/component/performance": {
      "get": {
        "summary": "Get Component Performance",
        "description": "Retrieve performance metrics for specific component(s)",
        "tags": ["Query"],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "component",
            "in": "query",
            "description": "Component/server name to filter by",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date_start",
            "in": "query",
            "description": "Start date",
            "required": false,
            "schema": {
              "type": "string",
              "default": "last_7_days"
            }
          },
          {
            "name": "date_end",
            "in": "query",
            "description": "End date",
            "required": false,
            "schema": {
              "type": "string",
              "default": "today"
            }
          },
          {
            "name": "metric",
            "in": "query",
            "description": "Specific metric to retrieve",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["all", "response_time", "avg_response_time", "max_response_time", "min_response_time", "success_rate", "error_rate"],
              "default": "all"
            }
          },
          {
            "name": "service_name",
            "in": "query",
            "description": "Filter by service name (partial match, case-insensitive)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "service_type",
            "in": "query",
            "description": "Filter by service type (exact match)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["API", "INTEGRATION", "AGENT", "SERVICE"]
            }
          },
          {
            "name": "statistic_name",
            "in": "query",
            "description": "Filter by statistic name (partial match, case-insensitive)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Performance metrics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerformanceResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/api/query/timeseries": {
      "get": {
        "summary": "Get Time Series Data",
        "description": "Retrieve time series data for visualization",
        "tags": ["Query"],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "component",
            "in": "query",
            "description": "Component name to filter by",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date_start",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "last_7_days"
            }
          },
          {
            "name": "date_end",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "today"
            }
          },
          {
            "name": "metric",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "response_time"
            }
          },
          {
            "name": "aggregation",
            "in": "query",
            "description": "Time aggregation level",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["hourly", "daily"],
              "default": "hourly"
            }
          },
          {
            "name": "service_name",
            "in": "query",
            "description": "Filter by service name (partial match, case-insensitive)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "service_type",
            "in": "query",
            "description": "Filter by service type (exact match)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["API", "INTEGRATION", "AGENT", "SERVICE"]
            }
          },
          {
            "name": "statistic_name",
            "in": "query",
            "description": "Filter by statistic name (partial match, case-insensitive)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Time series data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeSeriesResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/api/query/orders/stats": {
      "get": {
        "summary": "Get Order Statistics",
        "description": "Retrieve order statistics with aggregation",
        "tags": ["Query"],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "date_start",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "yesterday"
            }
          },
          {
            "name": "date_end",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "today"
            }
          },
          {
            "name": "statistic_type",
            "in": "query",
            "description": "Type of statistic to retrieve",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["maximum", "minimum"],
              "default": "maximum"
            }
          },
          {
            "name": "aggregation",
            "in": "query",
            "description": "Aggregation level",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["hourly", "daily", "weekly", "monthly"],
              "default": "daily"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderStatsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/api/query/components/compare": {
      "post": {
        "summary": "Compare Components",
        "description": "Compare performance metrics across multiple components",
        "tags": ["Query"],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["components"],
                "properties": {
                  "components": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": ["IKEAScheduleOrderIntegServer", "IkeaReleaseOrderAgent"]
                  },
                  "date_start": {
                    "type": "string",
                    "default": "last_7_days"
                  },
                  "date_end": {
                    "type": "string",
                    "default": "today"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comparison data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - components array required"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authentication"
      }
    },
    "schemas": {
...
      "StatsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "total_records": {
                "type": "string"
              },
              "unique_servers": {
                "type": "string"
              },
              "unique_services": {
                "type": "string"
              },
              "unique_statistics": {
                "type": "string"
              },
              "total_value": {
                "type": "string"
              },
              "avg_value": {
                "type": "string"
              },
              "max_value": {
                "type": "string"
              },
              "min_value": {
                "type": "string"
              },
              "earliest_record": {
                "type": "string",
                "format": "date-time"
              },
              "latest_record": {
                "type": "string",
                "format": "date-time"
              }
            }
          },
          "metadata": {
            "type": "object",
            "properties": {
              "dateStart": {
                "type": "string",
                "format": "date-time"
              },
              "dateEnd": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        }
      },
      "ComponentsListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "server_name": {
                  "type": "string"
                },
                "service_name": {
                  "type": "string"
                },
                "service_type": {
                  "type": "string"
                },
                "record_count": {
                  "type": "string"
                },
                "last_record_time": {
                  "type": "string",
                  "format": "date-time"
                },
                "statistic_types": {
                  "type": "string"
                }
              }
            }
          },
          "metadata": {
            "type": "object",
            "properties": {
              "searchTerm": {
                "type": "string"
              },
              "count": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              }
            }
          }
        }
      },
      "PerformanceResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "server_name": {
                  "type": "string"
                },
                "service_name": {
                  "type": "string"
                },
                "service_type": {
                  "type": "string"
                },
                "statistic_name": {
                  "type": "string"
                },
                "total_records": {
                  "type": "string"
                },
                "total_value": {
                  "type": "string"
                },
                "avg_value": {
                  "type": "string"
                },
                "max_value": {
                  "type": "string"
                },
                "min_value": {
                  "type": "string"
                }
              }
            }
          },
          "metadata": {
            "type": "object"
          }
        }
      },
      "TimeSeriesResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "time_period": {
                  "type": "string",
                  "format": "date-time"
                },
                "server_name": {
                  "type": "string"
                },
                "statistic_name": {
                  "type": "string"
                },
                "total_value": {
                  "type": "string"
                },
                "avg_value": {
                  "type": "string"
                },
                "max_value": {
                  "type": "string"
                },
                "min_value": {
                  "type": "string"
                },
                "record_count": {
                  "type": "string"
                }
              }
            }
          },
          "metadata": {
            "type": "object"
          }
        }
      },
      "OrderStatsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "nullable": true,
            "properties": {
              "time_period": {
                "type": "string",
                "format": "date-time"
              },
              "total_value": {
                "type": "string"
              },
              "record_count": {
                "type": "string"
              }
            }
          },
          "metadata": {
            "type": "object"
          }
        }
      },
      "ComparisonResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "metadata": {
            "type": "object",
            "properties": {
              "components": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "dateStart": {
                "type": "string"
              },
              "dateEnd": {
                "type": "string"
              },
              "count": {
                "type": "integer"
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "API key is missing or invalid",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}